services/web/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
click==7.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
email-validator==1.1.2
Flask-Assets==2.0
cssmin==0.2.0
jsmin==2.2.2
WTForms-SQLAlchemy==0.2